Simple Enum Migrations
They still aren't properly-defined in terms of Rust semantics, but it's a first step. Will come back later to work on things more in-depth. Today was a great first start at getting everything all set up for the next stages! Can't wait to see the rest of how this comes to fruition! This is already even better than what Region-Types on it's own could have been. Rust for the win! Using more than one language really is key here ⭐️