NBT Types Module
Looking into whether to use Rust newtypes vs aliases. I felt for the longest time that I wanted aliases, but that might not be the case afterall. This is a near copy of the definitions from my rust_nbt demo project.

Part of the goal here is that I want to constrain things like the ListTag generic to only allow NBT types.

I really like to have 'ambient mindfullness' while programming. It really helps my mind feel better. Sometimes I don't fully get the videos I find, but that's not a bad thing either. There are so many different styles out there, very interesting. This one sounds really great with my headphones, very relaxing. It really combines well with programming for me, because I can still get lost in the programming of things, but still be connected to reality, if that makes sense. I tend to get hyperfocused when I am in the zone for coding, and this really helps feel like I am still present, very great for that.

https://www.youtube.com/watch?v=Vq_Fx3YvTNQ