Tagged Tag Tags
Yes. This makes sense now, I promise.

So the thing is that there has to be a type that represents 'all' NBT types. That's what `Tag` is here. `TagLike` is what 'tags' whether the type is a valid type itself. So this is how the recursive part works, in reference to how NBTify handles things over in TypeScript land. I'm pretty sure this seemingly fixes the dilemma between things like CompoundTag and CompoundTagLike for example. Weird indexing issues. Yet to see though, of course!