Casing Revert
Realized I don't want to change these to match Rust specifics. Maybe later on, if I add a separate transform that defines version-specific casing, then I will store that separately. But for now at least, I want the enum to be unique to the real values that the game uses itself.

I think I haven't quite decided this still to some extent though, because with something like Alpha or Indev types for example, we have to 'make up' our own string IDs, since the game merely used plain number IDs back then. So I have yet to decide whether to follow a standard format across all editions, then merely have a custom transform that adjusts the casing/naming based on when it is serialized to the given real value, in a situation like NBT for example.