Allowed Custom Casing
You can use this flag above a declaration, or even just a single field, to use custom casing which the compiler usually warns about.

Because these are meant to specifically be internal representations of existing Minecraft IDs/names and such, they are kind of meant to be a bit off if that makes sense.

#[allow(non_camel_case_types)]

You can also make it a pragma which applies to the whole file, at the top of the file by adding an exclamation point.

#![allow(non_camel_case_types)]

There is one for allowing non-snake-case too. I will have to use that for struct fields as well, which need to mirror real NBT structures that already exist, and are not named consistently (unfortunately ha, hence the entire reason for needing to have this project in the first place).