Generic Tags!
Okay, I hope this works. This is exactly what I wanted things to work like in TS. But this is fully type-safe as well, not just a type annotation.

So in order to use an enum of values as an NBT value, you have to add an implementation to allow for using it. Awesome! So this still allows for third-party enums, just by adding their own impl blocks.

Will add these for the other types too.

Hoping that this is working as intended, we will see.

I wonder if I can still just use plain type aliases afterall too, or if I still need to use these newtype definitions? Will try that as well.