Type Alias NBT Types
Okay, so the key here is that yes, the type aliases do seem to work here. The key is that the specific list types that are distinct tags separate to that of a list tag of tags, those need to be newtypes because they would otherwise clash as being the same value. The compiler warns with an error in that case as well. Now it really does make sense in that regard.