Pi Entity Struct Composition Research
Went down the rabbit hole here.

https://www.rareskills.io/post/rust-attribute-derive-macro
https://stackoverflow.com/questions/32736170/what-is-the-best-way-to-inherit-a-struct-in-rust-1-3
https://www.reddit.com/r/learnrust/comments/o1vd9v/extending_struct_with_another/
https://www.reddit.com/r/learnrust/comments/1g30bsw/struct_inheritance_rust/
https://www.thecodedmessage.com/posts/oop-3-inheritance/
https://www.reddit.com/r/rust/comments/jzm4v6/how_to_add_fields_to_a_struct_via_attribute_macro/
https://www.reddit.com/r/rust/comments/18vbj19/rust_macro_for_generating_struct_fields_based_on/
https://crates.io/crates/derive_builder (interesting, a little different than what is needed for this specific road block)
https://users.rust-lang.org/t/using-a-macro-to-generate-struct-fields/53171/3 (closest vanilla implementation)
https://github.com/paulxtiseo/injectables (the solution!!!!)

Okay, so I know inheritance isn't a thing in Rust (lol here we go, right?)

I don't want 'regular inheritance'. I simply want a way to compose fields of structs without needing to write them a bunch.

I also need to mirror existing shapes which *are* modeled around inheritance, because these shapes come from languages which do have it. I also don't want to necessarily change the shape of the struct solely because I can't compose it without deduplication. I certainly *could* hypothetically just copy all of the fields to each of the instances. This is code-smelly though, and not maintainable. It is not an option I would want to take. So, why not compose repetitive fields with a macro?

That seems to be a great solution, to me. At least right now. I don't want to start introducing a bunch of crazy logic and libraries just because I can't represent the same structures flatly in the format I am trying to work with. Whether that's user-facing or not, that's another thing.

Maybe the goal here is afterall, that I will make these structures more Rust-friendly. That will be with time though. If I start doing that now, with no reference of usage or breakage, then it will just be totally useless. So I will refactor that once things make more sense to do so.