Further Pi Entity Types
Okay, almost there! Just have to add the rest of the composition macros in the next commit. I now really value committing code that is only easy to read through on first glance. That's the goal of commits, not for big sweeping changes. If it's not diffable, don't commit it!!!