Missing Entity Casing Flag
Okay, discovered why I am getting the recursive errors. They simply only seem to occur when you write code above the structs. So it's more of an editor-related thing, in that it's having to recompile them again, and it unsafely recompiles that code again, but the macro keeps running too, so it tries to assign it again, and it's already been done, so it doesn't like it. Not necessarily a bug, but not ideal I can see. It's fine for now though.