Big Java v1.20 Ports
Okay, not sure quite how long that was that I was working on things. But I got a heck of a lot done. So much that I don't want to risk losing this, even though I have more to do. Have to push it now, just in case. I did take a break midway through the entity one for a good summer cookout, and watching Poop Cruise and Tires. Wow, didn't realize how bad the cruise was. Not sure how we didn't hear about that when it happened. I feel like I kind of remember it, but not quite, at the same time. It was back in 2013.