Java Chunk and Structure Port
Almost there! Nearly all of the code is in Rust now. Wow, it's actually not *too* much, in a way. Definitely lots of room for improvement and refactoring. But this is the biggest necessary step to get to that point. This is already very awesome! Very happy with how relatively simple it was to move from TS to Rust. Now I want to try doing that for more projects of mine. To be fair, a lot of this was mostly around types though, so not much working logic to port. Glad I didn't write any of the working logic in TS.