Remove Java Registry Types
I think this was related to things running on the server. This project isn't specifically meant to handle things with that, so I am going to remove these types. Not that they aren't useful, but I don't personally have a use for these ones, so I don't want to try finagling it into something that works in Rust. This one looks especially difficult to translate too, since it uses a lot of key/value mapping at runtime (or rather compile time, for TypeScript).