Injectables Removal
Sad, this was a very nice implementation for things. It just didn't quite work out because of the issues with it being a macro. It can only work with a macro though, so it's not like there's another way to do this. If only the language itself supported 'mixins' for structs.

https://www.reddit.com/r/rust/comments/4kk5ir/composition_vs_inheritance_in_rust_tutorials/
https://www.artima.com/articles/composition-versus-inheritance

I think I am going to move forward by instead using what I have learned to be called Flattened Composition. I wanted to use ADT Composition, but it doesn't quite apply correctly here, because for example with entity structs, they borrow fields from more than one template, so they can't have a singular parent. Instead, I will have to pull in features with a container field itself, and handle flattening it with something like Serde.