Complete Injectables Removal
Moving over to Flattened Composition, keeping the original annotations as comments so I can migrate them in the next commits coming up. Discovering new usage errors now that the code is nearing no-error status. It's things like marking certain fields as public, I think I'll have to do it for essentially all of the structs, I don't know why I'd want them to be private for this. For a regular project I can see why, but this is specifically for defining all of the shapes of data which is meant to be edited.