Tile Tick File Rename
I think I am getting too into the swing of things, it is getting late, and I have been coding for a few hours. And a lot in general the last few days. Missing small details like this, making for more and more small commits. Maybe I am just worrying. It's okay, not a big deal. Just keep working while it feels right, then go to bed when it's time to rest :)

Listening to meditation ASMR again, it's really great.