Bedrock Block Module
I still have to do a lot of work to dedupe the types here, and also move over the `// extends` blocks. This one is kind of a big nightmare to work on, but absolutely worth it. It really is taxing on your system to run Rust Analyzer with it while editing it too, had to turn it off while working with most of this file. It at least appears to be error and warning free though, so I think we're good at the moment.

I can't believe the Rust version of this file is more than twice as many lines as the original TS version. Plain amazing. It will only get bigger with using flattened composition too, since I can't use 'extends' like you can in TS. Those can show on one line.