Bedrock Database Port
Heck yeah!

Bitten by the lossless CD bug. Not because it's "the best" in concept, but because it really does sound freaking fantastic. Going to have to double down on not re-buying a bunch of my existing iTunes buys as lossless. Special edition CD/DVD combos are super cool too, Boil That Dust Speck was an exceptionally outstanding buy. Now I want to get a bunch of Keneally's special edition releases. It makes me anxious that they are limited too, need to get them!!!! The video content is so freaking cool. Now I want to make things like that in the future too.