Schema Module Init
Going to start looking into adding conversion impl logic for all of the structs (wish me luck!). So I am moving the definitions themselves to a new module called schema. Adjacent to this will be the independent module, `convert` as it currently seems to want to be named, in my head at least.

Ana Maria - The Mark Varney Project