Convert: Java + Pi Setup
Going to try a 'simple' one to start with, since some of the library typings for Bedrock and LCE are a bit haywire still. Same with doing anything related to Java or Bedrock worlds themselves. That would involve already having chunk/region parsing working. Starting with the level.dat is a nice thing to see if it works properly.