More ConvertTo Demo Impls
Still have more to build out, this seems to be very maintainable though, love it! The whole tree of things here is very well type-checked, and very strongly linked together, while not needing to reference each other to work. That's the magic here. Rust is outstanding, love this language so much.