Pi Level -> Java (More Progress)
Coming back to the project after a few days, have to get my mind around how this works once again. More in just how I had organized things, the goal is to not repeat myself where it isn't necessary.

Please listen to Mike Keneally "Li'l (Signature Sound version)", I absolutely love that song. It is very high up there with "Cornbread Crumb" for me. Wow, I absolutely love his music. It's just so darn tangible! The room feel of it all, and that it's in the moment. It's liberating.