Java Pi Item Mapping
Okay, new scenario unfolding.

How does one handle item removal in the event of an ID not being present on the platform? The ID is an internal field in the Item itself, so how do I retroactively remove these if and when they are not available on that platform? Say for example, boats aren't available in Pi edition. Do I just return Air as the ID type? It sounds like that is an invalid item.

After thinking about it more, and asking GPT, it seems like the way to go is to make the `convert()` call an Option, and it can be handled from the top scope accordingly, the user of that method doesn't need to know how or why it failed, they just handle that it did (of course I would probably add error handling here of course, but in concept, you don't *need* to know why in order to use it).