Java Pi TryConvert + "All Mappings"!
Heck yeah! Okay, all of the item mappings are present here. I still have to figure out how to make this work for items that have block types too though, I didn't properly type that in the original project. And should it be it's own dedicated list, or simply an enum of block/item enums?